Add feed

+++ src/main/java/ru/yandex/practicum/filmorate/dao/FilmReviewStorage.java
@@ -1,11 +1,11 @@
 package ru.yandex.practicum.filmorate.dao;
 
 public interface FilmReviewStorage {
-    void addLikeToReview(Integer reviewId, Long userId);
+    void addLikeToReview(Long reviewId, Long userId);
 
-    void deleteLikeFromReview(Integer reviewId, Long userId);
+    void deleteLikeFromReview(Long reviewId, Long userId);
 
-    void addDislikeToReview(Integer reviewId, Long userId);
+    void addDislikeToReview(Long reviewId, Long userId);
 
-    void deleteDislikeFromReview(Integer reviewId, Long userId);
+    void deleteDislikeFromReview(Long reviewId, Long userId);
 }
